+++ src/clustering/naive.js
@@ -5,6 +5,9 @@
  * Naive clustering working by performing the n(n-1)/2 distance calculations
  * between all relevant pairs & retrieving connected components. Time complexity
  * of such a clustering is therefore O(n^2), which is quite bad.
+ *
+ * [Note]: this clusterer may create "chained" clusters where the diameter
+ * of the cluster is inintuively large.
  */
 
 /**