+++ src/regexp/classes.js
@@ -1,0 +1,8 @@
+/**
+ * Talisman regexp/classes
+ * ========================
+ *
+ * A collection of handy character classes.
+ */
+export const SINGLE_QUOTES = '’‘`‛\'';
+export const DOUBLE_QUOTES = '«»„‟“”"';