+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## 0.15.0
 
-* Adding `clustering/inverted-index` namespace.
+* Adding the `clustering/blocking` namespace.
+* Adding the `regexp` namespace.
 * Adding the `tokenizers/fingerprint` namespace.
 * Adding the `asymmetric` option to `clustering/naive`.
 * Adding the `minClusterSize` option to `clustering/key-collision`.
 * Adding the `minClusterSize` option to `clustering/naive`.
+* Adding the `minClusterSize` option to `clustering/vp-tree`.
 * Moving `stats/ngrams` to `tokenizers/ngrams`.
 * Moving `keyers/omission` to `phonetics/omission`.
 * Moving `keyers/skeleton` to `phonetics/skeleton`.